docmind/backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
pydantic-settings
langchain>=0.3
langchain-community
langchain-chroma
langchain-google-genai
langchain-experimental
langchain-huggingface
chromadb
sentence-transformers
unstructured[all-docs]
python-multipart
sse-starlette
rank_bm25
langchain-cohere
langchain-classic
